Refactor some things, make some changes
