+++ src/main/java/de/xtkq/voidgen/events/EventManager.java
@@ -1,29 +1,27 @@
 package de.xtkq.voidgen.events;
 
-import org.bukkit.plugin.java.JavaPlugin;
+import de.xtkq.voidgen.VoidGen;
 
 public class EventManager {
 
-    private final JavaPlugin javaPlugin;
+    private final VoidGen voidGen;
     private PlayerLoginListener playerLogin;
     private PluginEnableListener pluginEnable;
 
-    public EventManager(JavaPlugin paramPlugin) {
-        this.javaPlugin = paramPlugin;
+    public EventManager(VoidGen voidGen) {
+        this.voidGen = voidGen;
     }
 
     public void initialize() {
-        this.playerLogin = new PlayerLoginListener(this.javaPlugin);
-        this.pluginEnable = new PluginEnableListener(this.javaPlugin);
+        this.playerLogin = new PlayerLoginListener(this.voidGen);
+        this.pluginEnable = new PluginEnableListener(this.voidGen);
     }
 
     public void terminate() {
-        if (this.playerLogin != null) {
+        if (this.playerLogin != null)
             this.playerLogin.terminate();
-        }
-        if (this.pluginEnable != null) {
+        if (this.pluginEnable != null)
             this.pluginEnable.terminate();
-        }
     }
 }
