+++ src/main/java/de/xtkq/voidgen/events/PlayerLoginListener.java
@@ -1,5 +1,6 @@
 package de.xtkq.voidgen.events;
 
+import de.xtkq.voidgen.VoidGen;
 import de.xtkq.voidgen.utils.UpdateUtils;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
@@ -7,15 +8,14 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerLoginEvent;
-import org.bukkit.plugin.java.JavaPlugin;
 
 public class PlayerLoginListener implements Listener {
 
-    private final JavaPlugin plugin;
+    private final VoidGen voidGen;
 
-    public PlayerLoginListener(JavaPlugin paramPlugin) {
-        this.plugin = paramPlugin;
-        this.plugin.getServer().getPluginManager().registerEvents(this, plugin);
+    public PlayerLoginListener(VoidGen voidGen) {
+        this.voidGen = voidGen;
+        this.voidGen.getServer().getPluginManager().registerEvents(this, voidGen);
     }
 
     @EventHandler(priority = EventPriority.LOW)
@@ -23,17 +23,17 @@ public class PlayerLoginListener implements Listener {
         Player player = event.getPlayer();
         if (UpdateUtils.isUpdateAvailable()) {
             if (player.isOp()) {
-                this.plugin.getServer().getScheduler().runTaskLater(this.plugin, () -> player.sendMessage(this.getUpdateMessage()), 60L);
+                this.voidGen.getServer().getScheduler().runTaskLater(this.voidGen, () -> player.sendMessage(this.getUpdateMessage()), 60L);
             }
         }
     }
 
     private String getUpdateMessage() {
-        String updateMessage = String.format("&e%s &7v.%s is available here: &e%s&r", this.plugin.getName(), UpdateUtils.getLatestRelease(), UpdateUtils.getLatestReleaseURL());
+        String updateMessage = String.format("&e%s &7v.%s is available here: &e%s&r", this.voidGen.getName(), UpdateUtils.getLatestRelease(), UpdateUtils.getLatestReleaseURL());
         return ChatColor.translateAlternateColorCodes('&', updateMessage);
     }
 
     public void terminate() {
-        PlayerLoginEvent.getHandlerList().unregister(this.plugin);
+        PlayerLoginEvent.getHandlerList().unregister(this.voidGen);
     }
 }
