+++ src/main/java/de/xtkq/voidgen/events/PluginEnableListener.java
@@ -1,35 +1,36 @@
 package de.xtkq.voidgen.events;
 
-import de.xtkq.voidgen.hook.MultiverseGeneratorPluginHook;
+import de.xtkq.voidgen.VoidGen;
+import de.xtkq.voidgen.hooks.MultiverseGeneratorPluginHook;
 import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerLoginEvent;
 import org.bukkit.event.server.PluginEnableEvent;
-import org.bukkit.plugin.java.JavaPlugin;
 import org.mvplugins.multiverse.core.MultiverseCoreApi;
 
 public class PluginEnableListener implements Listener {
 
-    private final JavaPlugin plugin;
+    private final VoidGen voidGen;
 
-    public PluginEnableListener(JavaPlugin paramPlugin) {
-        this.plugin = paramPlugin;
-        this.plugin.getServer().getPluginManager().registerEvents(this, plugin);
+    public PluginEnableListener(VoidGen voidGen) {
+        this.voidGen = voidGen;
+        this.voidGen.getServer().getPluginManager().registerEvents(this, voidGen);
     }
 
-    @EventHandler
+    @EventHandler(priority = EventPriority.LOW)
     public void onPluginLoad(PluginEnableEvent event) {
         if (event.getPlugin().getName().equals("Multiverse-Core")) {
             try {
                 Class.forName("org.mvplugins.multiverse.core.MultiverseCoreApi");
                 MultiverseCoreApi.get().getGeneratorProvider().registerGeneratorPlugin(new MultiverseGeneratorPluginHook());
-            } catch (ClassNotFoundException e) {
+            } catch (ClassNotFoundException ignore) {
                 // ignore - assume multiverse not installed
             }
         }
     }
 
     public void terminate() {
-        PlayerLoginEvent.getHandlerList().unregister(this.plugin);
+        PlayerLoginEvent.getHandlerList().unregister(this.voidGen);
     }
 }
